+++ app/Policies/ExportPolicy.php
@@ -6,6 +6,7 @@ namespace App\Policies;
 
 use App\Models\User;
 use Filament\Actions\Exports\Models\Export;
+use Illuminate\Auth\Access\Response;
 
 /**
  * Defines the authorization rules for `Export` models.
@@ -18,27 +19,29 @@ use Filament\Actions\Exports\Models\Export;
  */
 final readonly class ExportPolicy
 {
-    /**
-     * Determines whether the given user can view the specified export.
-     *
-     * The policy grants permission only if the authenticated user is the owner of the export.
-     * It performs this check by comparing the user who initiated the export with the currently authenticated user.
-     *
-     * @param  User   $user    The authenticated user attempting to view the export.
-     * @param  Export $export  The export model instance being checked.
-     * @return bool            Returns `true` if the user is the owner of the export, otherwise `false`.
-     */
-    public function view(User $user, Export $export): bool
+	/**
+	 * Determines whether the given user can view the specified export.
+	 *
+	 * The policy grants permission only if the authenticated user is the owner of the export.
+	 * It performs this check by comparing the user who initiated the export with the currently authenticated user.
+	 *
+	 * @param  User   $user    The authenticated user attempting to view the export.
+	 * @param  Export $export  The export model instance being checked.
+	 * @return Response        Returns `true` if the user is the owner of the export, otherwise `false`.
+	 *
+	 * @throws \Exception
+	 */
+    public function view(User $user, Export $export): Response
     {
-        return $export->user()->is($user);
+        return ($export->user()->is($user))
+			? Response::allow()
+			: Response::deny();
     }
-
-    /**
-     * @todo Document policy
-     * @todo Implement policy on the action class
-     */
-    public function create(User $user): bool
+	
+    public function create(User $user): Response
     {
-        return $user->can('export_article');
+        return $user->can('export_article')
+			? Response::allow()
+			: Response::deny();
     }
 }